Completly move BCNumbersData to dedicated file
